feat: move create journey button in settings section
